Move Backwords +End
I have finished the script that allows the whale to move, both forwards
and backwords. I have also added a marioish (please don't read this and
sue us ditendo) flag to mark the end. There is one bug in the game that
desperately needs to be fixed, which is that the whale spins out of
control when ever it is uneaven. I believe this to be the reasult of 2D
physics, but either way it needs to be fixed.